feat(LottoCount): 수동으로 구매할 수 있는 로또의 수 validation checkt 기능 구현
