+++ README.md
@@ -106,6 +106,7 @@ utils.Calculator
 41. step4 구현해야 할 기능 설계
 42. 수동으로 숫자를 입력 받고, 그 결과를 출력하도록 기능 구현
 43. 사용하지 않는 code 제거
+44. 수동으로 구매할 수 있는 로또의 수 validation checkt 기능 구현
 
 
 ## 피드백 1
