+++ src/main/java/lotto/domain/Lottoes.java
@@ -14,16 +14,10 @@ public class Lottoes {
     private final List<LottoNumbers> autoLotto;
 
     public Lottoes(PurchaseAmount purchaseAmount, List<String> manualLotto) {
-        valid(purchaseAmount, manualLotto);
+
         this.totalAmount = purchaseAmount;
         this.manualLotto = new LottoNumberGenerator().create(manualLotto);
         this.autoLotto = new LottoNumberGenerator().create(getAutoLottoAmount());
-    }
-
-    private void valid(PurchaseAmount purchaseAmount, List<String> manualLotto) {
-        if (!purchaseAmount.canBuy(manualLotto.size())) {
-            throw new IllegalArgumentException(ErrorMessage.CAN_NOT_BUY_LOTTO);
-        }
     }
 
     private int getAutoLottoAmount() {
